Extract TimeView from drawing code
I also had to make a wrapper StackView here that... kind of does what I
wanted. What I really wanted was for it to center the clock along the X
and Y axis. What I did here at least made the clock visible, but the
wrapper stack view is hugging the text and not centering itself.