Refactor the TimeView to use a text field
I think this will be easier in the long run. All the text storage is
kind of a pain in the ass because you have to draw it at a specific
point and can't use auto layout as far as I know. This is much easier
to configure, use, and layout. Hopefully it will stay easy.