Center the time and hex code views
It seems like if I remove the top and height constraints, autolayout
can figure out the height of the wrapper view automatically from it's
subviews. But it still needs to know where to go vertically in the
view. So I added a constraint that centers the wrapper view in it's
superview.