Heavy refactor of the ColorGengerator
When I tried to make colors come out, they were 100% white. This is
because NSColor uses values from 0 - 1, not 0 - 255. So this is a
refactor to turn the base method `rgbComponentsFromTime` to return
values from 0 - 1. This unfortunately makes the `hexColorFromTime`
method a bit more complicated because I have to convert everything to 0
- 255 as Ints.