Load font from the local bundle
There were two problems fixed here:

First, I was loading the main bundle, which in this case is the
ScreenSaverEngine, not my screensaver. So I had to use my screensaver's
bundle to find the font.

Second, I was loading the font after it needed to be set in the views.
This was causing weird errors like one screen getting the font and
another having the default system font.