Animate background color
This actually might happen automatically if I just set the layer's
background color. It certainly looked like it was, but I don't know how
to adjust the timing of that yet.

Anyway, this stores the current background color and uses it to create
an animation that fades from one color to another over 1 second. I
tried setting it to 0.2 and it did transition much faster, but it kept
using the original color instead of the new color. I.e. instead of
fading from the currentBackground color, it would fade from the
original starting background color. Not sure why...