Cross Fade the whole thing
Well, this experiment was interesting.

I tried using CATransition to fade as it lets me set the state of the
view that I want and then transition to that state. Seems more
declarative in that way. The problem is, it cross faded _everything_
including the text. I'm going to try to see if I can get this to work
anyway.