Fade the animation without fading the text
I created a new view just for the background (BackgroundView) so it
could be separated from the transition. It works, but the
BackgroundView starts every time as transparent (which shows the plain
gray underneath) and fades to the start color. I tried to set a color
when the view is setup but it doesn't seem to take.