Move text view setups into configureViews()
This was originally an attempt to move the resizing of the fonts in
text views to each view's initialization process. While this worked in
the preview app, it didn't work at all in the actual screensaver. :(

It seemed that there was no superview for the text views to resize
from, so I couldn't set a size.