Animate textColor with an NSTimer
This is the first one that works the way I want it to. I have an
NSTimer run at 60 fps and set the textColor of the view from one color
to another. The exact shade is determined by the
`NSColor.blended(withFraction: ofColor:) method.

This is a pretty rough sketch of how it will work. I plan to refactor
in the next commit.