Refactor fading text labels
I created a new superclass called LabelView that both TimeView and
ColorCodeView inherit from. This takes care of initialization and fade
functionality.

I didn't want each view to constantly be calling fade, especially when
most of the time it's fading from one color to the same color. I keep
track of the `isLight` value of the background color. If that color
matches the current `isLight` value, then it won't call `fade`. If it
is different, the `fade` is called.

I went through many iterations of trying to detect if the color should
change or not. I ended up settling on this because it's the only thing
I could get to work. My original goal was to compare the view's
textColor with the potential new text color generated form the date
color. But comparing colors turns out to be difficult. Each color has a
color space and some color spaces don't like to be compared. Black and
white colors have one of these color spaces. Also, it seems that the
textColor on an NSTextField is converted into a color space that
doesn't like to be compared either.

All this made it challenging to use textColor in anyway as the check to
see if it should be faded to another color. So I kept track of a
boolean value instead and called it good enough.