Prevent image retention by making the text slightly transparent
Doing a bit of reading about image retention, it can be avoided by
letting the pixels change even a little bit. So I put a small bit of
transparency on the text color so those pixels (especially the colons)
would change. This combined with the black-to-white transitions should
remove most if not all possibility of screen retention.