Infrastructure for adding a 12-hour clock mode
This uses the UserDefaults system. It checks the value of the key
"isTwelveHour" which is a boolean value. There's a new class called
Settings which transforms the UserDefaults values, into actionable
values. For now, this includes only the appropriate date format based on
the "isTwelveHour" preference.

You may ask "Why not just store the format in the user defaults as a
string? Then you wouldn't need a wrapper class." The answer to this is
the user only knows 12-hour as a true/false value, so I wanted to mirror
that as much as possible in the code. The translation layer allows the
user (or the user specific code) to be dumb about date formatting, and
the TimeView to be dumb about user preferences.