Change dateFormat to an instance method
I realized that I need to use a different class for screen saver
settings (ScreenSaverDefaults) so this is prepping the code to be able
to change gracefully to that new class. ScreenSaverDefaults uses
instance methods. In order to store an instance, I need to keep an
instance of either that, or my Settings class around. So I'm switching
over to instance methods instead of class methods where it makes sense.