Add thumbnail for system preferences
This looks kind of crappy to be honest. The text is blurry and it
overlays an overlay on the image to try and match it up with other
screensaver thumbnails.

For the blurriness, I tried to make it the exact size of the image in
system preferences, but it still looked blurry. ¯\_(ツ)_/¯

As far as I know, there is no way to remove the gradient.