Fix screen saver preferences
When the preferences are updated, the changes must be synced back to
the system. This is especially true for screen savers, I believe,
because they are not attached to the same process that changes the
preferences. If you don't synchronize, you can get into a situation
where you see the preview change, but the screensaver it self does not.