Use referenced date instead of new date to calculate colors
I was doing a silly thing. When I was calling `asColor` or
`asHexColorString`, it would create a new date object and use that
date's time components (hour, minute, second) to create the color
values.

This worked for so long because I was always calling it on date object
that was the same date/time. But it doesn't work when you want to see
the color of some random date because it will only ever return the
current date's color, not the date you're calling it on.

This now uses `self` to refer to whatever date is calling those methods.